Revert "Component APIs and Life Cycles"
