+++ README.md
@@ -1,43 +1,1 @@
 # React-Starter-Kit
-
-1. Environment Setup
-2. JSX
-
-- Using JSX
-- Nested Elements
-- Javascript Expression
-- Ternary Expression
-
-3. Components
-
-- Stateless (Functional Component)
-- Statefull (Class Component)
-- Calling a Component
-
-4. State
-5. Props
-
-- Using Props
-- Default Props
-- State and Props
-- Props Validation
-- Validating Props
-
-6. Component APIs and Life Cycle
-
-- setState
-- Component Life Cycle
-
-7. Forms
-
-- Using Form
-- Complex Form
-
-8. Events
-
-- Using Events
-- Child Events
-
-9. Keys
-10. React Router
-11. Higher Order Components
