+++ package.json
@@ -3,12 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "node-sass": "4.13.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0"
+    "react-scripts": "3.2.0",
+    "node-sass": "4.13.0"
   },
   "scripts": {
     "start": "react-scripts start",
