+++ src/App.jsx
@@ -1,30 +1,46 @@
 import React from 'react';
-// import JSX from './Sample/JSX/usingJSX';   //to test other samples, change path
-// import Components from './Sample/Components/CallingAComponent';
-// import State from './Sample/State';
-// import Props from './Sample/Props/PropsValidation';
-// import ComponentAPIsAndLifeCycle from './Sample/ComponentAPIsAndLifeCycle/LifeCycles';
-// import Forms from './Sample/Forms/ComplexForm';
-// import Events from './Sample/Events/ChildEvents';
-// import Keys from './Sample/Keys';
-import HOC from './Sample/HOC';
+import PropTypes from 'prop-types';
 
 class App extends React.Component {
   render() {
     return (
       <div>
-        {/* <JSX/> */}
-        {/* <Components /> */}
-        {/* <State /> */}
-        {/* <Props /> */}
-        {/* <ComponentAPIsAndLifeCycle /> */}
-        {/* <Forms /> */}
-        {/* <Events /> */}
-        {/* <Keys /> */}
-        <HOC />
+        <h3>Array: {this.props.propArray}</h3>
+        <h3>Bool: {this.props.propBool ? 'True...' : 'False...'}</h3>
+        <h3>Func: {this.props.propFunc(3)}</h3>
+        <h3>Number: {this.props.propNumber}</h3>
+        <h3>String: {this.props.propString}</h3>
+        <h3>Object: {this.props.propObject.objectName1}</h3>
+        <h3>Object: {this.props.propObject.objectName2}</h3>
+        <h3>Object: {this.props.propObject.objectName3}</h3>
       </div>
     );
   }
 }
+
+App.propTypes = {
+  propArray: PropTypes.array.isRequired,
+  propBool: PropTypes.bool.isRequired,
+  propFunc: PropTypes.func,
+  propNumber: PropTypes.number,
+  propString: PropTypes.string,
+  propObject: PropTypes.object,
+};
+
+App.defaultProps = {
+  propArray: [1, 2, 3, 4, 5],
+  propBool: true,
+  propFunc: function(e) {
+    return e;
+  },
+  propNumber: 1,
+  propString: 'String value...',
+
+  propObject: {
+    objectName1: 'objectValue1',
+    objectName2: 'objectValue2',
+    objectName3: 'objectValue3',
+  },
+};
 
 export default App;
