+++ src/App.scss
@@ -1,3 +1,3 @@
-.test-app {
+.App {
   margin-left: 50px;
 }
