--- src/Routes/constant.routes.js
@@ -1,4 +1,0 @@
-export const LANDING = '/';
-export const HOME = '/home';
-export const ABOUT = '/about';
-export const CONTACT = '/contact';
