--- src/Sample/ComponentAPIsAndLifeCycle/LifeCycles/index.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import './style.scss';
-import Test from './Test';
-
-class LifeCycles extends React.Component {
-  render() {
-    return <Test />;
-  }
-}
-
-export default LifeCycles;
