--- src/Sample/ComponentAPIsAndLifeCycle/LifeCycles/style.scss
@@ -1,17 +1,0 @@
-.test-app {
-  margin-top: 50px;
-  margin-left: 100px;
-
-  .click-btn {
-    width: 150px;
-    height: 50px;
-    font-size: 20px;
-    margin-left: 20px;
-  }
-
-  .number {
-    margin-left: 167px;
-    margin-top: 35px;
-    font-size: 32px;
-  }
-}
