--- src/Sample/ComponentAPIsAndLifeCycle/SetState/style.scss
@@ -1,16 +1,0 @@
-.test-app {
-  margin-top: 50px;
-  margin-left: 100px;
-
-  .click-btn {
-    width: 150px;
-    height: 50px;
-    font-size: 20px;
-  }
-
-  .name {
-    margin-left: 50px;
-    font-size: 30px;
-    font-size: 30px;
-  }
-}
