--- src/Sample/Components/CallingAComponent/firstChild.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const FirstChild = () => {
-  return <p>I'm the 1st child!</p>;
-};
-
-export default FirstChild;
