--- src/Sample/Components/StatelessComponent/Content.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Content extends React.Component {
-  render() {
-    return (
-      <div>
-        <h2>Content</h2>
-        <p>The content text!!!</p>
-      </div>
-    );
-  }
-}
-
-export default Content;
