--- src/Sample/Components/StatelessComponent/Header.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class Header extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Header</h1>
-      </div>
-    );
-  }
-}
-
-export default Header;
