--- src/Sample/Events/ChildEvents/style.scss
@@ -1,16 +1,0 @@
-.test-app {
-  margin-top: 50px;
-  margin-left: 100px;
-
-  .click-btn {
-    width: 150px;
-    height: 50px;
-    font-size: 20px;
-  }
-
-  .text {
-    margin-left: 50px;
-    font-size: 30px;
-    font-size: 30px;
-  }
-}
