--- src/Sample/Forms/ComplexForm/style.scss
@@ -1,10 +1,0 @@
-.test-app {
-  margin: 70px;
-  font-size: 26px;
-  .input {
-    font-size: 26px;
-  }
-  .text {
-    margin-left: 15px;
-  }
-}
