--- src/Sample/JSX/javascriptExpression.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class JavascriptExpressions extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>{1 + 1}</h1>
-      </div>
-    );
-  }
-}
-
-export default JavascriptExpressions;
