--- src/Sample/JSX/nestedElements.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-class NestedElements extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Header</h1>
-        <h2>Content</h2>
-        <p>This is the content!!!</p>
-      </div>
-    );
-  }
-}
-
-export default NestedElements;
