--- src/Sample/JSX/styling.scss
@@ -1,3 +1,0 @@
-.header {
-  color: red;
-}
