--- src/Sample/JSX/usingJSX.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-class JSX extends React.Component {
-  render() {
-    return <div>Hello World!!!</div>;
-  }
-}
-
-export default JSX;
