--- src/Sample/Keys/Content.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Content extends React.Component {
-  render() {
-    return (
-      <div>
-        <div>{this.props.componentData.component}</div>
-        <div>{this.props.componentData.id}</div>
-      </div>
-    );
-  }
-}
-
-export default Content;
