--- src/Sample/Props/DefaultProps/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ParentComponent from './parentComponent';
-
-class DefaultProps extends React.Component {
-  render() {
-    return (
-      <div>
-        <ParentComponent />
-      </div>
-    );
-  }
-}
-
-export default DefaultProps;
