--- src/Sample/Props/UsingProps/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ParentComponent from './parentComponent';
-
-class UsingProps extends React.Component {
-  render() {
-    return (
-      <div>
-        <ParentComponent />
-      </div>
-    );
-  }
-}
-
-export default UsingProps;
