--- src/Sample/Props/UsingProps/parentComponent.jsx
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import FirstChild from './firstChild';
-
-class ParentComponent extends Component {
-  render() {
-    return (
-      <FirstChild
-        headerProp='Header from props...'
-        contentProp='Content from props...'
-      />
-    );
-  }
-}
-
-export default ParentComponent;
