--- src/Sample/Routes/About.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const About = () => (
-  <div>
-    <h1>About</h1>
-  </div>
-);
-
-export default About;
