--- src/Sample/Routes/Contact.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Contact = () => (
-  <div>
-    <h1>Contact</h1>
-  </div>
-);
-
-export default Contact;
