--- src/Sample/Routes/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Home = () => (
-  <div>
-    <h1>Home</h1>
-  </div>
-);
-
-export default Home;
