--- src/Sample/Routes/LandingPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const LandingPage = () => (
-  <div>
-    <h1>Landing Page</h1>
-  </div>
-);
-
-export default LandingPage;
